# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * E-Commerce Analytics & Recommendation System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type RegisterInputRole = typeof RegisterInputRole[keyof typeof RegisterInputRole];


export const RegisterInputRole = {
  admin: 'admin',
  customer: 'customer',
} as const;

export interface RegisterInput {
  name: string;
  email: string;
  password: string;
  role?: RegisterInputRole;
}

export interface User {
  id: number;
  name: string;
  email: string;
  role: string;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface Product {
  id: number;
  name: string;
  category: string;
  price: number;
  stock: number;
  rating: number;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  createdAt?: string;
}

export interface ProductInput {
  name: string;
  category: string;
  price: number;
  stock: number;
  rating?: number;
  description?: string;
  imageUrl?: string;
}

export interface ProductUpdate {
  name?: string;
  category?: string;
  price?: number;
  stock?: number;
  rating?: number;
  description?: string;
  imageUrl?: string;
}

export interface ProductList {
  items: Product[];
  total: number;
  page: number;
  limit: number;
}

export interface ProductSalesStat {
  id: number;
  name: string;
  category: string;
  totalSold: number;
  revenue: number;
  rating: number;
}

export interface CategoryPerformance {
  category: string;
  totalSales: number;
  revenue: number;
  productCount: number;
  avgRating: number;
}

export interface Customer {
  id: number;
  name: string;
  email: string;
  totalOrders: number;
  totalSpent: number;
  segment: string;
  clv?: number;
  /** @nullable */
  lastOrderDate?: string | null;
  createdAt: string;
}

export interface OrderItem {
  id: number;
  productId: number;
  productName: string;
  quantity: number;
  price: number;
}

export interface Order {
  id: number;
  userId: number;
  /** @nullable */
  customerName?: string | null;
  totalAmount: number;
  orderDate: string;
  status?: string;
  items: OrderItem[];
}

export interface RFMScore {
  recency: number;
  frequency: number;
  monetary: number;
  total: number;
}

export interface CustomerDetail {
  id: number;
  name: string;
  email: string;
  totalOrders: number;
  totalSpent: number;
  segment: string;
  clv?: number;
  /** @nullable */
  lastOrderDate?: string | null;
  createdAt: string;
  recentOrders: Order[];
  rfmScore: RFMScore;
}

export interface CustomerList {
  items: Customer[];
  total: number;
  page: number;
  limit: number;
}

export interface CustomerSegment {
  segment: string;
  count: number;
  avgClv: number;
  avgOrderValue: number;
  percentage: number;
}

export interface RetentionMetrics {
  retentionRate: number;
  churnRate: number;
  avgPurchaseFrequency: number;
  avgDaysBetweenOrders: number;
  repeatCustomerRate: number;
}

export interface OrderItemInput {
  productId: number;
  quantity: number;
}

export interface OrderInput {
  userId: number;
  items: OrderItemInput[];
}

export interface OrderList {
  items: Order[];
  total: number;
  page: number;
  limit: number;
}

export interface AnalyticsSummary {
  totalRevenue: number;
  totalOrders: number;
  totalCustomers: number;
  totalProducts: number;
  avgOrderValue: number;
  revenueGrowth: number;
  ordersGrowth: number;
  customersGrowth: number;
}

export interface RevenueDataPoint {
  date: string;
  revenue: number;
  orders: number;
}

export interface SalesReport {
  period: string;
  revenue: number;
  orders: number;
  avgOrderValue: number;
  topProduct: string;
}

export interface CategoryProfit {
  category: string;
  revenue: number;
  cost: number;
  profit: number;
  margin: number;
}

export interface ProfitAnalysis {
  totalRevenue: number;
  estimatedCost: number;
  grossProfit: number;
  grossMargin: number;
  byCategory: CategoryProfit[];
}

export interface ActivityItem {
  id: number;
  type: string;
  description: string;
  /** @nullable */
  amount: number | null;
  /** @nullable */
  customerName?: string | null;
  timestamp: string;
}

export interface ProductRecommendation {
  id: number;
  name: string;
  category: string;
  price: number;
  rating: number;
  score: number;
  reason: string;
  /** @nullable */
  imageUrl?: string | null;
}

export interface ForecastPoint {
  date: string;
  predicted: number;
  lower: number;
  upper: number;
  /** @nullable */
  actual?: number | null;
}

export interface InventoryForecast {
  productId: number;
  productName: string;
  currentStock: number;
  predictedDemand: number;
  reorderPoint: number;
  /** @nullable */
  daysUntilStockout: number | null;
  recommendation: string;
}

export type ListProductsParams = {
category?: string;
search?: string;
page?: number;
limit?: number;
};

export type GetTopSellingProductsParams = {
limit?: number;
};

export type ListCustomersParams = {
segment?: string;
search?: string;
page?: number;
limit?: number;
};

export type ListOrdersParams = {
page?: number;
limit?: number;
from?: string;
to?: string;
};

export type GetRevenueAnalyticsParams = {
period?: GetRevenueAnalyticsPeriod;
from?: string;
to?: string;
};

export type GetRevenueAnalyticsPeriod = typeof GetRevenueAnalyticsPeriod[keyof typeof GetRevenueAnalyticsPeriod];


export const GetRevenueAnalyticsPeriod = {
  daily: 'daily',
  weekly: 'weekly',
  monthly: 'monthly',
} as const;

export type GetSalesReportParams = {
period?: GetSalesReportPeriod;
};

export type GetSalesReportPeriod = typeof GetSalesReportPeriod[keyof typeof GetSalesReportPeriod];


export const GetSalesReportPeriod = {
  daily: 'daily',
  monthly: 'monthly',
} as const;

export type GetRecentActivityParams = {
limit?: number;
};

export type GetTrendingProductsParams = {
limit?: number;
};

export type GetSalesForecastParams = {
periods?: number;
granularity?: GetSalesForecastGranularity;
};

export type GetSalesForecastGranularity = typeof GetSalesForecastGranularity[keyof typeof GetSalesForecastGranularity];


export const GetSalesForecastGranularity = {
  daily: 'daily',
  weekly: 'weekly',
  monthly: 'monthly',
} as const;