# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: E-Commerce Analytics & Recommendation System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /analytics/revenue:
    get:
      operationId: getRevenueAnalytics
      tags: [analytics]
      summary: Revenue over time
      parameters:
        - in: query
          name: period
          schema:
            type: string
            enum: [daily, weekly, monthly]
        - in: query
          name: from
          schema:
            type: string
        - in: query
          name: to
          schema:
            type: string
      responses:
        "200":
          description: Revenue data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RevenueDataPoint"
  /analytics/sales-report:
    get:
      operationId: getSalesReport
      tags: [analytics]
      summary: Detailed sales report
      parameters:
        - in: query
          name: period
          schema:
            type: string
            enum: [daily, monthly]
      responses:
        "200":
          description: Sales report
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SalesReport"
...
  /analytics/recent-activity:
    get:
      operationId: getRecentActivity
      tags: [analytics]
      summary: Recent order activity feed
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Recent activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"

  # RECOMMENDATIONS
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    RegisterInput:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [admin, customer]
    User:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        createdAt:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
...
    RevenueDataPoint:
      type: object
      required: [date, revenue, orders]
      properties:
        date:
          type: string
        revenue:
          type: number
        orders:
          type: integer
    SalesReport:
      type: object
      required: [period, revenue, orders, avgOrderValue, topProduct]
      properties:
        period:
          type: string
        revenue:
          type: number
        orders:
          type: integer
        avgOrderValue:
          type: number
        topProduct:
          type: string
...
    ActivityItem:
      type: object
      required: [id, type, description, amount, timestamp]
      properties:
        id:
          type: integer
        type:
          type: string
        description:
          type: string
        amount:
          type: ["number", "null"]
        customerName:
          type: ["string", "null"]
        timestamp:
          type: string
...